Buffer: replace INIT and INIT_RAW with constructors
One of the reason we use INIT instead of constructor was out-of-memory
(OOM) handling. We thought that by using INIT we can check if OOM happened
the return value would be NULL. That is true, so does the `new Buffer`.

So, there is no reason why we should differentiate INIT and INIT_RAW with
class constructor since both do the same things while the later was much
simpler.

We also remove INIT_SIZE, because its like calling Buffer constructor
with integer.