Replace Object string representation attribute from `_v` to `__str`
We thought that we can use `_v` so other object that inherited from
class Buffer does not have duplicate attribute.

Turn out we are wrong. We must differentiate between "what" the object is
and what the object content is. For example, a server have buffer `_v` to
read and write to client. When we want to know the information about
server, by printing it with `chars()`, it should not return the content
of buffer but other information, e.g. IP address and port that they bind
to.