make: allow user to use different compiler and linker
Previously, we set the compiler (CXX) and linker (LINKER) default to
`gcc`. In this commit we check if user define the CXX and LINKER and use
them.

This commit remove `-Werror`. Since we only check if program can be
compiled and build on `gcc`, the `-Werror` option can cause compilation
to stop if there is warning when using other compiler (e.g. when using
clang).

This commit also remove link-time-optimizer when building program because
its not compatible with `-Os`.