Resolver: do not check for reply code, keep going, we are not police
Previously, we check the reply code (RCODE) from answer packet and return
it as fail if RCODE is not zero. This will cause some program "hang"
because they wait for reply but none is received, so they request again
and again.

For example, when querying non existent DNS name, like "thisisnotexist",
the RCODE would contain 3. Chrome browser said they use this to prevent
ISP hijacking the response with their own search page.